@@ -11,71 +11,71 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.0",
-    "@angular/cdk": "^7.1.0",
-    "@angular/common": "^6.1.0",
-    "@angular/compiler": "^6.1.0",
-    "@angular/core": "^6.1.0",
-    "@angular/forms": "^6.1.0",
-    "@angular/http": "^6.1.0",
-    "@angular/material": "^7.1.1",
-    "@angular/platform-browser": "^6.1.0",
-    "@angular/platform-browser-dynamic": "^6.1.0",
-    "@angular/router": "^6.1.0",
-    "@ngrx/effects": "^6.1.2",
-    "@ngrx/entity": "^6.1.2",
-    "@ngrx/store": "^6.1.2",
-    "@ngrx/store-devtools": "^6.1.2",
-    "algoliasearch": "^3.30.0",
-    "bootstrap": "^3.3.7",
-    "bootstrap-material-design": "^0.5.10",
-    "bootstrap-sass": "^3.3.7",
+    "@angular/animations": "^7.2.5",
+    "@angular/cdk": "^7.3.2",
+    "@angular/common": "^7.2.5",
+    "@angular/compiler": "^7.2.5",
+    "@angular/core": "^7.2.5",
+    "@angular/forms": "^7.2.5",
+    "@angular/http": "^7.2.5",
+    "@angular/material": "^7.3.2",
+    "@angular/platform-browser": "^7.2.5",
+    "@angular/platform-browser-dynamic": "^7.2.5",
+    "@angular/router": "^7.2.5",
+    "@ngrx/effects": "^7.2.0",
+    "@ngrx/entity": "^7.2.0",
+    "@ngrx/store": "^7.2.0",
+    "@ngrx/store-devtools": "^7.2.0",
+    "algoliasearch": "^3.32.0",
+    "bootstrap": "^4.3.1",
     "chartist": "^0.11.0",
-    "core-js": "^2.5.4",
+    "core-js": "^2.6.5",
     "cropit": "^0.5.1",
     "enquire.js": "^2.1.6",
     "eonasdan-bootstrap-datetimepicker": "^4.17.47",
-    "flv.js": "^1.4.2",
+    "flv.js": "^1.5.0",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
-    "howler": "^2.0.15",
+    "howler": "^2.1.1",
     "jquery": "^3.3.1",
     "material-datetime-picker": "^2.4.0",
-    "moment": "^2.20.1",
-    "ng2-bs3-modal": "^0.15.0",
+    "moment": "^2.24.0",
     "ng2-img-cropper": "^0.9.0",
-    "ngx-infinite-scroll": "^6.0.1",
-    "normalize.css": "^7.0.0",
+    "ngrx-store-localstorage": "^7.0.0",
+    "ngx-infinite-scroll": "^7.0.1",
+    "normalize.css": "^8.0.1",
     "prismic-dom": "^2.1.0",
     "prismic-javascript": "^2.0.1",
     "raven-js": "^3.23.3",
-    "rxjs": "^6.0.0",
-    "snackbarjs": "^1.1.0",
-    "socket.io-client": "^2.0.3",
+    "rxjs": "^6.4.0",
+    "socket.io-client": "^2.2.0",
     "store": "^2.0.12",
-    "sweetalert2": "^6.6.5",
-    "zone.js": "~0.8.26"
+    "sweetalert2": "^8.2.1",
+    "tslib": "^1.9.0",
+    "zone.js": "~0.8.29"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.7.0",
-    "@angular/cli": "~6.1.4",
-    "@angular/compiler-cli": "^6.1.0",
-    "@angular/language-service": "^6.1.0",
-    "@ngrx/schematics": "^6.1.2",
-    "@types/jasmine": "^2.8.12",
+    "@angular-devkit/build-angular": "~0.13.0",
+    "@angular/cli": "~7.3.2",
+    "@angular/compiler-cli": "^7.2.5",
+    "@angular/language-service": "^7.2.5",
+    "@ngrx/schematics": "^7.2.0",
+    "@types/jasmine": "^3.3.9",
     "@types/jasminewd2": "^2.0.6",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
-    "jasmine-core": "~2.99.1",
+    "@types/node": "~11.9.4",
+    "@types/stripe": "^6.19.7",
+    "@types/stripe-v3": "^3.0.9",
+    "codelyzer": "~4.5.0",
+    "jasmine-core": "~3.3.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
+    "karma": "~4.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.0",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.0",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "~2.7.2"
+    "karma-coverage-istanbul-reporter": "~2.0.5",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "protractor": "~5.4.2",
+    "ts-node": "~8.0.2",
+    "tslint": "~5.12.1",
+    "typescript": "^3.2.4"
   }
 }
