@@ -1,6 +1,7 @@
-import { MatSnackBar } from '@angular/material';
+import { MatDialog, MatSnackBar } from '@angular/material';
+import { PromptModalComponent } from 'app/shared/prompt-modal.component';
 import { User } from 'models/user';
-import { first, map, share, switchMap, tap } from 'rxjs/operators';
+import { filter, first, map, share, switchMap, tap } from 'rxjs/operators';
 import { Component, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
 import * as moment from 'moment';
@@ -32,6 +33,7 @@ export class SessionsComponent implements OnInit {
     private userService: UserService,
     private toastService: ToastService,
     private snack: MatSnackBar,
+    private dialog: MatDialog,
   ) {
   }
 
@@ -39,7 +41,6 @@ export class SessionsComponent implements OnInit {
     const bookings$ = this.auth.getCurrentUser(true).pipe(
       first(),
       switchMap((u: User) => this.bookableService.getBookingsForUser(u._id)),
-      tap((b) => console.log(b)),
       share(),
     );
 
@@ -55,31 +56,43 @@ export class SessionsComponent implements OnInit {
 
   cancel(session: Booking) {
 
-    this.modal.popup({
-      type: 'info',
-      html: `Are you sure you want to cancel your session on ${moment(session.time.start).format('MMM Do @ h:mm a')}? This action cannot be undone. *** Due to full schedules, if you choose to cancel your session, you will only receive 50% of your payment. If your session is within 72 hours, there will be no refund given. There will be a grace period of 1 hour after booking.`,
-      confirmButtonText: 'Cancel Session',
-      cancelButtonText: 'Nevermind',
-    })
-      .then(doCancel => {
-        if (!doCancel) return;
+    const dialogRef = this.dialog.open(PromptModalComponent, {
+      data: {
+        title: 'Cancel Session',
+        body: [
+          `Are you sure you want to cancel your session on ${moment(session.time.start).format('MMM Do @ h:mm a')}? This action cannot be undone.`,
+          `*** Due to full schedules, if you choose to cancel your session, you will only receive 50% of your payment. If your session is within 72 hours, there will be no refund given. There will be a grace period of 1 hour after booking.`
+        ],
+        confirmButtonText: 'Cancel Session',
+        confirmButtonColor: 'warn',
+      },
+      width: '500px',
+    });
 
-        this.bookableService.cancel(session)
-          .subscribe(
-            res => {
-              Object.assign(session, res);
-              this.snack.open('Session cancelled successfully', null, {
-                duration: 5000,
-                horizontalPosition: 'right',
-                verticalPosition: 'top',
-              });
-            },
-            err => {
-              console.log(err);
-              this.modal.error();
-            }
-          )
-      })
+    dialogRef.beforeClosed()
+      .pipe(
+        filter(confirm => !!confirm),
+        switchMap(() => this.bookableService.cancel(session)),
+      )
+      .subscribe(
+        res => {
+          Object.assign(session, res);
+          this.snack.open('Session cancelled successfully', null, {
+            duration: 5000,
+            horizontalPosition: 'right',
+            verticalPosition: 'top',
+          });
+        },
+        err => {
+          console.log(err);
+          this.snack.open('Something went wrong', null, {
+            duration: 5000,
+            horizontalPosition: 'right',
+            verticalPosition: 'top',
+            panelClass: 'snack-error',
+          })
+        }
+      );
 
   }
 }
