@@ -1,3 +1,4 @@
+import { FormBuilder, FormGroup, Validators as val } from '@angular/forms';
 import { first } from 'rxjs/operators';
 import { AfterViewInit, Component, ErrorHandler, OnDestroy, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
@@ -11,61 +12,52 @@ import { ModalService } from 'app/core/modal.service';
   templateUrl: './login.component.html',
   styleUrls: ['./login.component.scss'],
 })
-export class LoginComponent implements OnInit, OnDestroy, AfterViewInit {
-  errorMessage: string;
+export class LoginComponent implements OnInit {
   isLoading: boolean;
-
-  subscriptions: any = {};
+  loginForm: FormGroup;
 
   constructor(
     private auth: AuthService,
     private router: Router,
     private modal: ModalService,
     private http: HttpClient,
-    private errorReporter: ErrorHandler
+    private errorReporter: ErrorHandler,
+    private fb: FormBuilder,
   ) {
+    this.loginForm = fb.group({
+      email: ['', [val.required, val.email]],
+      password: ['', [val.required]]
+    });
   }
 
   ngOnInit() {
-    this.auth.isLoggedIn().pipe(
-      first())
-      .subscribe(is => {
-        if (is) this.router.navigate(['']);
-      });
+    this.auth.isLoggedIn().pipe(first())
+      .subscribe(is => is && this.router.navigate(['']));
+
+    this.loginForm.valueChanges
+      .subscribe(() => {
+        const { errors } = this.loginForm.get('email');
+        if (!errors) return;
+        delete errors.server;
+        if (Object.keys(errors).length === 0) this.loginForm.get('email').setErrors(null);
+      })
   }
 
-  ngAfterViewInit() {
-    $.material.init('home-login *');
-    setTimeout(() => {
-      const input = (<any>document.querySelector('home-login input[autofocus]'));
-      if (input) input.focus();
-    }, 1000);
-  }
-
-  ngOnDestroy() {
-    if (this.subscriptions.login) this.subscriptions.login.unsubscribe();
-  }
-
-  onSubmit(form) {
-
+  onSubmit(form: FormGroup) {
     if (form.invalid) return;
     this.isLoading = true;
 
-    this.subscriptions.login = this.auth.login(form.value.email, form.value.password)
+    this.auth.login(form.value.email, form.value.password)
       .subscribe(
         () => {
-          this.errorMessage = null;
-          this.router.navigate(['']);
+          return this.router.navigate(['']);
         },
         err => {
-          this.isLoading = false;
-          if (err.statusText === 'Unauthorized') {
-            this.errorMessage = 'Email or password was incorrect.';
-          } else {
-            this.errorMessage = 'Something went wrong.';
-            this.errorReporter.handleError(err);
-          }
-        }
+          this.loginForm.get('email').setErrors({
+            server: err.statusText === 'Unauthorized' ? 'Email or password was incorrect.' : 'Something went wrong.',
+          })
+        },
+        () => this.isLoading = false,
       );
   }
 
@@ -75,7 +67,6 @@ export class LoginComponent implements OnInit, OnDestroy, AfterViewInit {
       text: 'Enter your email and we\'ll email you a link to reset your password',
       input: 'email',
       inputPlaceholder: 'Email',
-      closeOnConfirm: false,
     })
       .then(email => {
         if (!email) return;
